# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Fashion Xpress API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: registerCustomer
      tags: [auth]
      summary: Register a new customer account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSession"
        "400":
          description: Invalid input
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login as customer, admin, or executive
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSession"
        "401":
          description: Invalid credentials
  /auth/me:
    get:
      operationId: getCurrentUser
      tags: [auth]
      summary: Get currently authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated

  # ---------------- CUSTOMER PROFILE ----------------
...
components:
...
  schemas:
...
    User:
      type: object
      required: [id, email, name, role, createdAt]
      properties:
        id: { type: integer }
        email: { type: string }
        name: { type: string }
        phone: { type: ["string", "null"] }
        role: { type: string, enum: [customer, admin, executive] }
        createdAt: { type: string, format: date-time }
    RegisterInput:
      type: object
      required: [name, email, password, phone]
      properties:
        name: { type: string, minLength: 1 }
        email: { type: string }
        password: { type: string, minLength: 6 }
        phone: { type: string, minLength: 1 }
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email: { type: string }
        password: { type: string }
    AuthSession:
      type: object
      required: [token, user]
      properties:
        token: { type: string }
        user:
          $ref: "#/components/schemas/User"

    # ---------- CUSTOMER ----------
...