# lib/api-zod/src/generated/types/authSession.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Fashion Xpress API specification
 * OpenAPI spec version: 0.1.0
 */
import type { User } from './user';

export interface AuthSession {
  token: string;
  user: User;
}